+++ src/MyBot/TranspositionTable.cs
@@ -26,7 +26,7 @@ public class TranspositionTable
         table = new TranspositionData[size];
     }
 
-    public void Put(Board board, byte depth, byte flag, int eval, Move? bestMove)
+    public void Put(Board board, byte depth, byte flag, double eval, Move? bestMove)
     {
         // int index = (int) (board.ZobristHash % (ulong)size);
         int index = TableIndex(board.ZobristHash);
